Various docs fixes (#18)
This PR collects various docs fixes that I'm drive by implementing as I
read the docs for the first time. @JoelOtter I will let you know when
this is ready for review.

**Update: This PR is now ready for review. It includes all fixes
necessary to make this setup flow function correctly on macOS. These
are:**

- Correct the path to `test_app`
- Correct the `./growl-cmd assets` to take into account the fact that
`assets` is two directories back, not one.
- Add make as a dependency.
- Link to the homebrews for make and cmake to help users find
dependencies and push them towards using the homebrew workflow that we
use.